Fixed throwing exception while printing shortcut labels

+++ lib/shortcut.py
@@ -315,7 +315,7 @@ class Shortcuts(QtCore.QObject):
             icon_width = scaled_icon.width()
             icon_height = scaled_icon.height()
 
-            painter.drawImage(x + (label_width - icon_width) / 2, y + (label_height - icon_height) / 2, scaled_icon)
+            painter.drawImage(int(x + (label_width - icon_width) / 2), int(y + (label_height - icon_height) / 2), scaled_icon)
             painter.setPen(QtGui.QColor("black"))
             painter.drawRect(x, y, label_width, label_height)
